requirements.txt
requests>=2.31.0
python-telegram-bot[job-scheduler]>=22.0
python-dotenv>=1.0.0
openai>=1.0.0
beautifulsoup4>=4.12.0
